Codebase/requirements.txt
torch
transformers
accelerate
peft
soundfile
python-dotenv
kokoro
